// apps/web/src/types/UpdateNode.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NodeType } from "./NodeType";

export type UpdateNode = { text?: string, semantic_id?: string, node_type?: NodeType, display_category?: string, position_x?: number, position_y?: number, is_active?: boolean, };